Add index.html into /public
Now we can see our project file structure starts to evolve. We don’t
want any of our source files be accessible to users, so we make a
dedicated directory for everything, that is going to be served by web
server. And we name it “public” so it reminds it’s purpose do
developers.