Add webpack
It’s convenient to separate your source code into files and
directories, but you don’t want users’ browsers to make a separate
request for every of hundreds of your modules.

Webpack has a consensus for you here. It will concatenate and minify
your source files into a bundle, so users can download scripts for your
web app in a single request.

And there are many cool other things webpack has to offer! We are going
to see that in later commits.